+++ 2019_edition/advent-of-java/src/main/java/day02/instructions/IntCodeInstructionAddition.java
@@ -7,14 +7,14 @@ import day02.IntCodeComputer;
 public class IntCodeInstructionAddition extends IntCodeInstructionAbstract
 	implements IntCodeInstruction {
 
-    public IntCodeInstructionAddition(int instructionCode) {
+    public IntCodeInstructionAddition(Long instructionCode) {
         super(instructionCode, 3);
     }
 
     @Override
-    public void apply(IntCodeComputer computer, List<Integer> parameters) {
-        Integer writeIndex = parameters.get(2);
-        Integer result = parameters.get(0) + parameters.get(1);
+    public void apply(IntCodeComputer computer, List<Long> parameters) {
+        Long writeIndex = parameters.get(2);
+        Long result = parameters.get(0) + parameters.get(1);
         computer.setMemoryAddress(writeIndex, result);
     }
