+++ 2019_edition/advent-of-java/src/main/java/day02/instructions/IntCodeInstructionHalt.java
@@ -6,12 +6,12 @@ import day02.IntCodeComputer;
 
 public class IntCodeInstructionHalt extends IntCodeInstructionAbstract {
 
-    public IntCodeInstructionHalt(int opCode) {
+    public IntCodeInstructionHalt(Long opCode) {
         super(opCode, 0);
     }
 
     @Override
-    public void apply(IntCodeComputer computer, List<Integer> parameters) {
+    public void apply(IntCodeComputer computer, List<Long> parameters) {
         computer.requestHalt("Program execution complete");
     }
