+++ 2019_edition/advent-of-java/src/main/java/day02/instructions/IntCodeInstructionMultiplication.java
@@ -6,14 +6,14 @@ import day02.IntCodeComputer;
 
 public class IntCodeInstructionMultiplication extends IntCodeInstructionAbstract {
 
-	public IntCodeInstructionMultiplication(int opCode) {
+	public IntCodeInstructionMultiplication(Long opCode) {
 		super(opCode, 3);
 	}
 
 	@Override
-	public void apply(IntCodeComputer computer, List<Integer> parameters) {
-		Integer writeIndex = parameters.get(2);
-        Integer result = parameters.get(0) * parameters.get(1);
+	public void apply(IntCodeComputer computer, List<Long> parameters) {
+		Long writeIndex = parameters.get(2);
+        Long result = parameters.get(0) * parameters.get(1);
         computer.setMemoryAddress(writeIndex, result);
 	}
