+++ 2019_edition/advent-of-java/src/main/java/day07/instructions/IntCodeInstructionWithMissingInput.java
@@ -14,7 +14,7 @@ public abstract class IntCodeInstructionWithMissingInput extends IntCodeInstruct
 	}
 
     @Override
-    public void apply(IntCodeComputer computer, List<Integer> parameters) {
+    public void apply(IntCodeComputer computer, List<Long> parameters) {
     	IntCodeComputer7 computer7 = (IntCodeComputer7) computer;
     	String instructionName = this.getClass().getSimpleName();
 		System.out.printf("Instruction %s requires input: \n", instructionName);
